Hacking on improved output. Have reference matching done.
Adjusted resolution rules so that symbol tokens with $'s in their name first attempt to resolve globally before stepping through the scoped resolution rules.

I made this adjustment so that label references could be rewritten while individual modules are being parsed/uplifted. That is happening now. So, the -l option
should provide a properly globalized listing of the code.